basic functioning app with configuration options
This initial commit of functioning code shows a basic
template for analyzing and graphing expenses.  I've
replaced my own variables with CHANGEME notes to help
others customize this to fit their own budgeting needs.

Some of the widgets are currently blank because I have't
finished implementing them yet.

The code is also a little fragile with respect to formatting
types and data restrictions.